Add type model to store opportunity types
The JSON output of ODF recreational opportunities contains a "type"
string for each record (opportunity). The string is a comma-separated
list of stuff to do at that location.

I eventually want to store each value within each comma-separated list
as its own type, and do a many-to-many relationship between
opportunities and types. So this is probably getting ahead of things,
but go ahead and create a basic Type model for this eventual
information.