Drop dependency on rails_12factor
I've been annoyed by lots of chatter from Postgres in my command line. I
had assumed it was a configurable Postgres thing, but turns out that
it's a byproduct of the rails_12factor gem that provided Heroku support
in older versions of Rails. Since this is a Rails 5 app, it's no longer
needed.

Drop the dependency on rails_12factor to clean up console output.