Import opportunities' types
Each recreational opportunity has one or more "types" that we can
already store in the database, but we don't have a way to associate them
with the actual opportunities.

Add a join table to associate opportunities with these types, then use
that to finish the import tool by breaking up the "type" key provided in
the ODF JSON file into individual terms (it's a comma-delimited string),
then adding each. We might be able to optimize this someday, but for
now, with such a small data set, I'm just looking for forward progress.