Add Google Analytics code
I want to be able to see if this project gains any traction going
forward. Google Analytics is free and well-documented, so I'm going with
that for now.

Add the analytics code and store it as an encrypted credential. Even
though it renders in plain text in page source, using encrypted
credentials is the recommended way to store environment-specific configs
in Rails 5.2 and beyond.

Turbolinks tip from this post:
https://michaelsoolee.com/google-analytics-rails-5/