Add link to directions for each opportunity
I eventually want to do more with the location data made available in
the data set, but a quick win is linking off to Google Maps for
directions.

Add a link to get directions from each opportunity's card in the list,
pointing to Google Maps, using the latitude and longitude provided in
the data set.

Doing this work, I realized I reversed latitude and longitude when
importing data. Since I've already got live data in production, I'll fix
this in a followup.